Handles null case (#103)
