+++ app/components/listing/index.tsx
@@ -6,7 +6,7 @@ export * from "./OrganizationCard";
 export * from "./OrganizationDetails";
 export * from "./RelativeOpeningTime";
 export * from "../ui/Cards/ServiceCard";
-export * from "./ServiceDetails";
+export * from "./ServiceAttribution";
 export * from "./StreetViewImage";
 export * from "./ContactInfoTableRows";
 export * from "./TableOfLocationTimes";
